+++ packages/ethereum-contracts/contracts/agreements/InstantDistributionAgreementV1.sol
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: AGPLv3
 pragma solidity ^0.8.23;
 
+import { CallbackUtils } from "../libs/CallbackUtils.sol";
 import { SafeCast } from "@openzeppelin-v5/contracts/utils/math/SafeCast.sol";
 
 import {
@@ -372,6 +373,7 @@ contract InstantDistributionAgreementV1 is
          IndexData idata;
          SubscriptionData sdata;
          bytes cbdata;
+         uint256 remainingCallbackGas;
      }
 
     /// @dev IInstantDistributionAgreementV1.approveSubscription implementation
@@ -419,7 +421,8 @@ contract InstantDistributionAgreementV1 is
 
         if (!vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_CREATED_NOOP;
-            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
 
             vars.sdata = SubscriptionData({
                 publisher: publisher,
@@ -433,10 +436,12 @@ contract InstantDistributionAgreementV1 is
             token.createAgreement(vars.sId, _encodeSubscriptionData(vars.sdata));
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_CREATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
             // NOTE casting these values to int256 is okay because the original values
             // are uint128
             int balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -456,7 +461,8 @@ contract InstantDistributionAgreementV1 is
             token.updateAgreementData(vars.sId, _encodeSubscriptionData(vars.sdata));
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
         }
 
         // can index up to three words, hence splitting into two events from publisher or subscriber's view.
@@ -505,7 +511,8 @@ contract InstantDistributionAgreementV1 is
         newCtx = ctx;
 
         cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
-        (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+        (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+            cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
         // NOTE downcasting these values to int256 is okay because the original values
         // are uint128
         int256 balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -527,7 +534,8 @@ contract InstantDistributionAgreementV1 is
         token.settleBalance(subscriber, balanceDelta);
 
         cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
-        (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+        (, newCtx) = AgreementLibrary.callAppAfterCallback(
+            cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
 
         emit IndexUnsubscribed(token, publisher, indexId, subscriber, userData);
         emit SubscriptionRevoked(token, subscriber, publisher, indexId, userData);
@@ -576,10 +584,12 @@ contract InstantDistributionAgreementV1 is
         // before-hook callback
         if (vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_CREATED_NOOP;
-            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
         }
 
         // update publisher data
@@ -643,10 +653,12 @@ contract InstantDistributionAgreementV1 is
         // after-hook callback
         if (vars.subscriptionExists) {
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
         } else {
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_CREATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
         }
 
         emit IndexUnitsUpdated(token, publisher, indexId, subscriber, units, userData);
@@ -799,7 +811,8 @@ contract InstantDistributionAgreementV1 is
         newCtx = ctx;
 
         cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_TERMINATED_NOOP;
-        (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+        (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+            cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
         // NOTE casting these values to int256 is okay because the original values
         // are uint128
         int256 balanceDelta = int256(uint256(vars.idata.indexValue - vars.sdata.indexValue))
@@ -831,7 +844,8 @@ contract InstantDistributionAgreementV1 is
         token.settleBalance(subscriber, balanceDelta);
 
         cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_TERMINATED_NOOP;
-        (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+        (, newCtx) = AgreementLibrary.callAppAfterCallback(
+            cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
 
         emit IndexUnsubscribed(token, publisher, indexId, subscriber, userData);
         emit SubscriptionRevoked(token, subscriber, publisher, indexId, userData);
@@ -882,7 +896,8 @@ contract InstantDistributionAgreementV1 is
 
         if (pendingDistribution > 0) {
             cbStates.noopBit = SuperAppDefinitions.BEFORE_AGREEMENT_UPDATED_NOOP;
-            (vars.cbdata, newCtx) = AgreementLibrary.callAppBeforeCallback(cbStates, newCtx);
+            (vars.cbdata, vars.remainingCallbackGas) = AgreementLibrary.callAppBeforeCallback(
+                cbStates, CallbackUtils.HOST_CALLBACK_GAS_LIMIT, newCtx);
             int256 signedPendingDistribution = pendingDistribution.toInt256();
 
             // adjust publisher's deposits
@@ -898,7 +913,8 @@ contract InstantDistributionAgreementV1 is
             emit SubscriptionDistributionClaimed(token, subscriber, publisher, indexId, pendingDistribution);
 
             cbStates.noopBit = SuperAppDefinitions.AFTER_AGREEMENT_UPDATED_NOOP;
-            (, newCtx) = AgreementLibrary.callAppAfterCallback(cbStates, vars.cbdata, newCtx);
+            (, newCtx) = AgreementLibrary.callAppAfterCallback(
+                cbStates, vars.cbdata, vars.remainingCallbackGas, newCtx);
         } else {
             // nothing to be recorded in this case
             newCtx = ctx;