+++ packages/ethereum-contracts/test/contracts/superfluid/Superfluid.test.ts
@@ -762,14 +762,18 @@ describe("Superfluid Host Contract", function () {
                         superfluid.callAppBeforeCallback(
                             ZERO_ADDRESS,
                             "0x",
-                            "0x"
+                            false,
+                            "0x",
+                            0
                         )
                     );
                     await expectReverted(
                         superfluid.callAppAfterCallback(
                             ZERO_ADDRESS,
                             "0x",
-                            "0x"
+                            false,
+                            "0x",
+                            0
                         )
                     );
                     await expectReverted(
@@ -781,9 +785,7 @@ describe("Superfluid Host Contract", function () {
                             ZERO_ADDRESS
                         )
                     );
-                    await expectReverted(
-                        superfluid.appCallbackPop("0x", 0, "0x")
-                    );
+                    await expectReverted(superfluid.appCallbackPop("0x", 0));
                     await expectReverted(superfluid.ctxUseCredit("0x", 0));
                 });